Separate per-file actions from treewalk boilerplate
