+++ file-list.txt
@@ -1,0 +1,10 @@
+/root/foo_OLD	1536	FS_FILE_TYPE_DIRECTORY
+/root/bar	512	FS_FILE_TYPE_DIRECTORY
+/root/.DS_Store	6148	FS_FILE_TYPE_FILE
+/root/foo	1024	FS_FILE_TYPE_DIRECTORY
+/root/foo_OLD/afile.txt	40	FS_FILE_TYPE_FILE
+/root/foo_OLD/anotherfile.txt	49	FS_FILE_TYPE_FILE
+/root/foo_OLD/.DS_Store	6148	FS_FILE_TYPE_FILE
+/root/bar/qumulo-2.14.0.2.qimg	1246392320	FS_FILE_TYPE_FILE
+/root/foo/yetanotherfile.txt	62	FS_FILE_TYPE_FILE
+/root/foo/.DS_Store	6148	FS_FILE_TYPE_FILE
